+++ src/vue/filters/formatPercent.js
@@ -1,5 +1,0 @@
-import { globalize } from './globalize'
-
-export function formatPercent (value) {
-  return globalize('formats.percent', { value: value / 100 })
-}